// ws-client/src/app/school-erp/common/shared/model/institute/institute-class-section.ts
export class InstituteClassSection {

    oid: string;
    nameEn: string;
    nameBn: string;
    instituteOid: string;
    instituteSessionOid: string;
    instituteClassOid: string;
    instituteClassGroupOid: string;
    instituteShiftOid: string;
    instituteVersionOid: string;
    status: string;

}