// ws-client/src/app/school-erp/common/shared/model/subject/class-subject-list.ts
export class ClassSubjectList {

  oid: string;
  nameEn: string;
  nameBn: string;
  subjectCode: string;
  subjectType: string;
  status: string;
  instituteOid: string;
  instituteSessionOid: string;
  instituteVersionOid: string;
  instituteClassGroupOid: string;
  instituteClassOid: string;
  educationSubjectOid: string;
  subjectNameEn: string;
  subjectNameBn: string;

  //used in service
  check: boolean;
}